backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Message broker
aio-pika==9.3.1

# WebSocket
websockets==12.0

# HTTP client
aiohttp==3.9.1

# ML and data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
shap==0.43.0

# Monitoring
prometheus-client==0.19.0

# Bybit API
pybit==5.7.0

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1